+++ index.ts
@@ -2,3 +2,4 @@ export { NextCloudApi } from './credentials/NextCloudApi.credentials';
 export { NextCloud } from './nodes/NextCloud/NextCloud.node';
 export { NextCloudSpreadsheet } from './nodes/NextCloudSpreadsheet/NextCloudSpreadsheet.node';
 export { NextCloudDocTemplate } from './nodes/NextCloudDocTemplate/NextCloudDocTemplate.node';
+export { NextCloudPdf } from './nodes/NextCloudPdf/NextCloudPdf.node';